Target yaw was limited to 256 by 8 bits, changed to 16

+++ sw/peripherals-board/src/drivers/hwrev2/hwrev2_single_lidar_open_round.hpp
@@ -39,7 +39,7 @@ private:
 
     // About IMU
     float yaw;
-    uint8_t targetYaw = 0;
+    uint16_t targetYaw = 0;
     int startYaw = 0;
 
     unsigned long startMillis;
